requirements.txt
#Fill in dependencies here
#Fill in dependencies here
django
djangorestframework==3.9.3
psycopg2-binary
gunicorn
drf-yasg
pyyaml
